# admin-backend/docs/swagger.yaml
definitions:
  admin_proto.AdminApiItem:
    properties:
      createdAt:
        type: string
      describe:
        type: string
      id:
        type: integer
      isEnabled:
        type: boolean
      key:
        type: string
      name:
        type: string
      path:
        type: string
      permissionId:
        type: integer
      updatedAt:
        type: string
    type: object
  admin_proto.AdminInfo:
    properties:
      adminId:
        type: integer
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      expire:
        type: integer
      expireDataTime:
        type: string
      lastLoginIp:
        type: string
      lastLoginTime:
        type: string
      loginTotal:
        type: integer
      menus:
        additionalProperties:
          $ref: '#/definitions/admin_proto.MenuItem'
        type: object
      nickname:
        type: string
      permissions:
        additionalProperties:
          type: string
        type: object
      roles:
        items:
          $ref: '#/definitions/admin_proto.AdminUserRoleItem'
        type: array
      token:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
...
  admin_proto.AdminUserRoleItem:
    properties:
      roleId:
        description: 角色ID
        type: integer
      roleName:
        type: string
    type: object
  admin_proto.MenuItem:
    properties:
      authority:
        type: string
      component:
        type: string
      hideChildrenInMenu:
        type: boolean
      hideInMenu:
        type: boolean
      icon:
        description: 菜单图标
        type: string
      key:
        description: 菜单唯一键
        type: string
      name:
        description: 菜单名称
        type: string
      path:
        description: 菜单路由
        type: string
      routes:
        items:
          $ref: '#/definitions/admin_proto.MenuItem'
        type: array
    type: object
...
  admin_proto.RespAdminApiListData:
    properties:
      list:
        items:
          $ref: '#/definitions/admin_proto.AdminApiItem'
        type: array
      total:
        type: integer
    type: object
...
  admin_proto.RespLoginData:
    properties:
      data:
        $ref: '#/definitions/admin_proto.AdminInfo'
    type: object
...
  code.Message:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/code_proto.ErrorCode'
        description: 错误码
      data: {}
      msg:
        description: 错误码描述
        type: string
      reason:
        description: 错误具体信息
        type: string
    type: object
  code_proto.ErrorCode:
    enum:
    - 0
    - 1
    - 100001
    - 100002
    - 100003
    - 200001
    - 200002
    - 200003
    - 200004
    - 200005
    - 300001
    - 300002
    - 300003
    - 400001
    - 400002
    - 400003
    - 500001
    - 500002
    - 500003
    - 500004
    - 600001
    - 600002
    - 600003
    - 600004
    - 600005
    - 700001
    - 800001
    type: integer
    x-enum-comments:
      ErrorCode_AdminAccountInvalid: 账号已被封禁
      ErrorCode_AdminAccountNotExist: 账号不存在或已被删除
      ErrorCode_AdminAccountPasswordInvalid: 账号或密码错误
      ErrorCode_AdminApiKeyExist: 接口键名已存在
      ErrorCode_AdminApiNameExist: 接口名称已存在
      ErrorCode_AdminApiNotExist: 接口不存在货已失效
      ErrorCode_AdminApiPathExist: 接口路径已存在
      ErrorCode_AdminMenuNotExist: 菜单不存在
      ErrorCode_AdminPermissionExist: 该菜单下权限已存在
      ErrorCode_AdminPermissionKeyExist: 权限键名已存在
      ErrorCode_AdminPermissionKeyNeed: 权限键名不能为空
      ErrorCode_AdminPermissionNameNeed: 权限名称不能为空
      ErrorCode_AdminPermissionTypeInvalid: 权限类型无效
      ErrorCode_AdminRoleNotExist: 角色不存在
      ErrorCode_AuthTokenFailed: 鉴权失败
      ErrorCode_AuthTokenForbidden: 没有访问权限
      ErrorCode_AuthTokenInfoInvalid: Token信息无效
      ErrorCode_AuthTokenInspectInvalid: BearerToken无效
      ErrorCode_AuthTokenInvalid: BearerToken无效
      ErrorCode_Error: 失败
      ErrorCode_ReadContextRequestBodyFailed: 获取请求体内容失败
      ErrorCode_RecordNValidCanNotDeleted: 生效中的记录不能删除
      ErrorCode_RecordNotExist: 记录不存在
      ErrorCode_RequestBodyInvalid: 请求体参数无效
      ErrorCode_RequestParamsInvalid: 请求参数无效
      ErrorCode_RequestQueryInvalid: 请求查询参数无效
      ErrorCode_Success: 成功
    x-enum-varnames:
    - ErrorCode_Success
    - ErrorCode_Error
    - ErrorCode_ReadContextRequestBodyFailed
    - ErrorCode_RecordNotExist
    - ErrorCode_RecordNValidCanNotDeleted
    - ErrorCode_AuthTokenFailed
    - ErrorCode_AuthTokenInvalid
    - ErrorCode_AuthTokenInspectInvalid
    - ErrorCode_AuthTokenInfoInvalid
    - ErrorCode_AuthTokenForbidden
    - ErrorCode_RequestBodyInvalid
    - ErrorCode_RequestQueryInvalid
    - ErrorCode_RequestParamsInvalid
    - ErrorCode_AdminAccountPasswordInvalid
    - ErrorCode_AdminAccountNotExist
    - ErrorCode_AdminAccountInvalid
    - ErrorCode_AdminApiNameExist
    - ErrorCode_AdminApiPathExist
    - ErrorCode_AdminApiKeyExist
    - ErrorCode_AdminApiNotExist
    - ErrorCode_AdminPermissionKeyExist
    - ErrorCode_AdminPermissionExist
    - ErrorCode_AdminPermissionKeyNeed
    - ErrorCode_AdminPermissionNameNeed
    - ErrorCode_AdminPermissionTypeInvalid
    - ErrorCode_AdminMenuNotExist
    - ErrorCode_AdminRoleNotExist
info:
  contact: {}
paths:
  /admin/account/edit:
    post:
      consumes:
      - application/json
      description: 编辑账号
      parameters:
      - description: '头像 @gotags:  binding:"url" label:"头像地址"'
        in: query
        name: avatar
        type: string
      - description: '邮箱 @gotags:  binding:"email" label:"邮箱"'
        in: query
        name: email
        type: string
      - description: '名称 @gotags:  binding:"required,min=1,max=32" label:"用户名"'
        in: query
        maxLength: 32
        minLength: 1
        name: nickname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 编辑账号
      tags:
      - 账号相关接口
...
  /admin/account/login:
    post:
      consumes:
      - application/json
      description: 登录后台
      parameters:
      - description: '密码  @gotags: binding:"required,adminpwd" label:"密码"'
        in: query
        name: password
        required: true
        type: string
      - description: '用户名 @gotags: binding:"required,adminname" label:"用户名"'
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  $ref: '#/definitions/admin_proto.RespLoginData'
              type: object
      summary: 登录后台
      tags:
      - 账号相关接口
...
  /admin/account/password:
    post:
      consumes:
      - application/json
      description: 修改密码
      parameters:
      - description: '确认密码  @gotags: binding:"required,adminpwd,eqfield=Password"
          label:"确认密码"'
        in: query
        name: confirmPassword
        required: true
        type: string
      - description: '旧密码 @gotags:  binding:"required,adminpwd" label:"旧密码"'
        in: query
        name: oldPassword
        required: true
        type: string
      - description: '新密码 @gotags:  binding:"required,adminpwd" label:"新密码"'
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 修改密码
      tags:
      - 账号相关接口
...
  /admin/api/add:
    post:
      consumes:
      - application/json
      description: 新增接口
      parameters:
      - description: '接口描述 @gotags:  label:"接口描述" binding:"required,trimBlank"'
        in: query
        name: describe
        required: true
        type: string
      - description: 启用状态
        in: query
        name: enabled
        type: boolean
      - description: '接口键名 @gotags:  label:"接口键名" binding:"required,adminApiKey"'
        in: query
        name: key
        required: true
        type: string
      - description: '接口名称 @gotags:  label:"接口名称" binding:"required,trimBlank"'
        in: query
        name: name
        required: true
        type: string
      - description: '接口路由 @gotags:  label:"接口路由" binding:"required,adminApiPath"'
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 新增接口
      tags:
      - API接口相关
...
  /admin/api/delete:
    post:
      consumes:
      - application/json
      description: 接口删除
      parameters:
      - description: '接口ID @gotags:  label:"接口ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 接口删除
      tags:
      - API接口相关
  /admin/api/edit:
    post:
      consumes:
      - application/json
      description: 接口编辑
      parameters:
      - description: '接口描述 @gotags:  label:"接口描述" binding:"required,trimBlank"'
        in: query
        name: describe
        required: true
        type: string
      - description: 接口状态
        in: query
        name: enabled
        type: boolean
      - description: '接口ID @gotags:  label:"接口ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      - description: '接口键名 @gotags:  label:"接口键名" binding:"required,adminApiKey"'
        in: query
        name: key
        required: true
        type: string
      - description: '接口名称 @gotags:  label:"接口名称" binding:"required,trimBlank"'
        in: query
        name: name
        required: true
        type: string
      - description: '接口路由 @gotags:  label:"接口路由" binding:"required,adminApiPath"'
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 接口编辑
      tags:
      - API接口相关
  /admin/api/enable:
    post:
      consumes:
      - application/json
      description: 接口启用/禁用
      parameters:
      - description: 接口状态
        in: query
        name: enabled
        type: boolean
      - description: '接口ID @gotags:  label:"接口ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  type: object
              type: object
      summary: 接口启用/禁用
      tags:
      - API接口相关
  /admin/api/info:
    post:
      consumes:
      - application/json
      description: 接口详情
      parameters:
      - description: '接口ID @gotags:  label:"接口ID" binding:"required,min=1"'
        in: query
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  $ref: '#/definitions/admin_proto.AdminApiItem'
              type: object
      summary: 接口详情
      tags:
      - API接口相关
  /admin/api/list:
    post:
      consumes:
      - application/json
      description: 接口列表
      parameters:
      - description: 接口键名
        in: query
        name: key
        type: string
      - description: 接口名称
        in: query
        name: name
        type: string
      - description: 接口路由
        in: query
        name: path
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/code.Message'
            - properties:
                data:
                  $ref: '#/definitions/admin_proto.RespAdminApiListData'
              type: object
      summary: 接口列表
      tags:
      - API接口相关
...
swagger: "2.0"